move tabBar into own component
