+++ packages/apolloschurchapp/src/tabs/index.js
@@ -1,32 +1,10 @@
-import React from 'react';
 import { createBottomTabNavigator } from 'react-navigation';
-import { BottomTabBar } from 'react-navigation-tabs';
-import { Query } from 'react-apollo';
-import gql from 'graphql-tag';
+
+import TabBar from './tabBar';
 
 import Connect from './connect';
 import Home from './home';
 import Discover from './discover';
-
-const mediaPlayerIsVisible = gql`
-  query {
-    mediaPlayer @client {
-      isVisible
-    }
-  }
-`;
-const TabBarComponent = (props) => (
-  <Query query={mediaPlayerIsVisible}>
-    {({ data = {} }) => {
-      const overrideProps = {};
-      if (data.mediaPlayer && data.mediaPlayer.isVisible) {
-        overrideProps.safeAreaInset = { bottom: 'never', top: 'never' };
-      }
-      console.log({ overrideProps });
-      return <BottomTabBar {...props} {...overrideProps} />;
-    }}
-  </Query>
-);
 
 const TabNavigator = createBottomTabNavigator(
   {
@@ -35,7 +13,7 @@ const TabNavigator = createBottomTabNavigator(
     Connect,
   },
   {
-    tabBarComponent: TabBarComponent,
+    tabBarComponent: TabBar,
     tabBarOptions: {
       showLabel: false,
       activeTintColor: '#17B582', // TODO: get from theme
